+++ CHANGELOG.md
@@ -16,6 +16,7 @@ To see unreleased changes, please see the [CHANGELOG on the master branch](https
 
 * mTLS support.
 * `expose.mtls_ca_cert` setting.
+* `destroy` command.
 
 ## 0.9.1 - 2025-06-24